requirements.txt
beautifulsoup4==4.11.1
certifi==2022.9.24
charset-normalizer==2.1.1
distlib==0.3.6
ExifRead==3.0.0
filelock==3.8.0
idna==3.4
imagesize==1.4.1
pew==1.2.0
Pillow==9.3.0
platformdirs==2.5.2
requests==2.28.1
shellingham==1.5.0
soupsieve==2.3.2.post1
urllib3==1.26.13
virtualenv==20.16.5
virtualenv-clone==0.5.7
